% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_counter.R
\name{blob_counter}
\alias{blob_counter}
\title{blobcounter}
\usage{
blob_counter(
  mask_file,
  tracks_file,
  corner_thres = 120,
  centroid_similarity_thres = 10
)
}
\arguments{
\item{mask_file}{the mask file used with trackR}

\item{tracks_file}{the trajectory csv output file from trackR}

\item{corner_thres}{how sensitive the corner detection is}

\item{centroid_similarity_thres}{how similar centroids can be to each other without being discarded}
}
\value{
a data frame
}
\description{
blobcounter
}
\examples{
mask_file_input <- "C:/Users/imura/Downloads/7.19(B1)60D_mask2_1.png"
tracks_file_input <- 'C:/Users/imura/Downloads/sanity_tracks4.csv'
blob_counter(mask_file_input, tracks_file_input)
}
