% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction_blob_counter.R
\name{blob_direction}
\alias{blob_direction}
\title{blob_direction}
\usage{
blob_direction(tracks_file)
}
\arguments{
\item{tracks_file}{the trajectory csv output file from trackR}
}
\value{
a data frame
}
\description{
blob_direction
}
\examples{
tracks_file_input <- 'C:/Users/imura/Downloads/sanity_tracks4.csv'
blob_direction(tracks_file_input)
}
